add babel polyfill for async await, move entry and output config to webpack file
