+++ client/src/actions/types.js
@@ -1,0 +1,2 @@
+export const AUTH_USER = 'auth_user';
+export const AUTH_ERROR = 'auth_error';
