+++ package.json
@@ -7,8 +7,8 @@
     "start": "node server/index.js",
     "dev": "nodemon server/index.js",
     "lint": "eslint ./",
-    "react-dev": "webpack --mode development --watch ./client/src/index.jsx --output ./client/dist/bundle.js",
-    "react-build": "webpack --mode production ./client/src/index.jsx --output ./client/dist/bundle.js",
+    "react-dev": "webpack --mode development --watch",
+    "react-build": "webpack --mode production",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -27,19 +27,19 @@
     "url": "https://github.com/Postaff/Postaff/issues"
   },
   "dependencies": {
-    "@material-ui/core": "^1.2.3",
+    "@material-ui/core": "^1.3.0",
     "@material-ui/icons": "^1.1.0",
     "apollo-boost": "^0.1.10",
     "apollo-client": "^2.3.5",
     "apollo-server": "^2.0.0-rc.4",
     "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
     "bcrypt": "^2.0.1",
+    "body-parser": "^1.18.3",
     "chart.js": "^2.7.2",
     "classnames": "^2.2.6",
     "connect-flash": "^0.1.1",
-    "cors": "^2.8.4",
     "cookie-parser": "^1.4.3",
+    "cors": "^2.8.4",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "express-session": "^1.15.6",
@@ -48,18 +48,22 @@
     "graphql-tools": "^3.0.4",
     "helmet": "^3.12.1",
     "i": "^0.3.6",
+    "jwt-simple": "^0.5.1",
     "lodash": "^4.17.10",
     "material-ui-rating": "^2.0.0",
     "morgan": "^1.9.0",
     "mysql2": "^1.5.3",
-    "jwt-simple": "^0.5.1",
     "passport": "^0.4.0",
-    "passport-local": "^1.0.0",
     "passport-jwt": "^4.0.0",
+    "passport-local": "^1.0.0",
     "react": "^16.4.1",
     "react-apollo": "^2.1.6",
     "react-dom": "^16.4.1",
+    "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
+    "redux": "^4.0.0",
+    "redux-form": "^7.4.2",
+    "redux-thunk": "^2.3.0",
     "sequelize": "^4.38.0",
     "styled-components": "^3.3.3"
   },
@@ -75,6 +79,7 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
+    "babel-polyfill": "6.26.0",
     "css-loader": "^0.28.11",
     "eslint": "^5.0.1",
     "eslint-config-airbnb": "^17.0.0",
